Small fixes for mouse positioning. Clean-up for win32 CD reader code.

+++ extras/audio plugins/wrapper/AU/juce_AU_Wrapper.mm
@@ -794,7 +794,7 @@ public:
 
                 while (i.getNextEvent (midiEventData, midiEventSize, midiEventPosition))
                 {
-                    jassert (isPositiveAndBelow (midiEventPosition, numSamples));
+                    jassert (isPositiveAndBelow (midiEventPosition, (int) numSamples));
